Improve translation bookkeeping control, auto-purging
We can still keep translated formulae around by passing in
`keep_translation = true`.

+++ src/API.jl
@@ -382,19 +382,24 @@ function deps_sorted(pkg::BrewPkg; build_deps::Bool=false)
 end
 
 """
-`add(pkg::Union{AbstractString,BrewPkg}; verbose::Bool=false)`
+`add(pkg::Union{AbstractString,BrewPkg}; verbose::Bool=false, keep_translations::Bool=false)`
 
 Install package `pkg` and all dependencies, using bottles only, unlinking any
 previous versions if necessary, and linking the new ones in place. Will attempt
 to install non-relocatable bottles from `Homebrew/core` by translating formulae
 and forcing `cellar :any` into the formulae.
-"""
-function add(pkg::StringOrPkg; verbose::Bool=false) end
 
-function add(name::AbstractString; verbose::Bool=false)
+Automatically deletes all translated formulae before adding formulae and after,
+unless `keep_translations` is set to `true`.
+"""
+function add(pkg::StringOrPkg; verbose::Bool=false, keep_translations::Bool=false) end
+
+function add(name::AbstractString; verbose::Bool=false, keep_translations::Bool=false)
     # We do this so that things are as unambiguous and fresh as possible.
     # It's not a bad situation because translation is very fast.
-    delete_all_translated_formulae()
+    if !keep_translations
+        delete_all_translated_formulae()
+    end
 
     # Begin by translating all dependencies of `name`, including :build deps.
     # We need to translate :build deps so that we don't run into the diamond of death
@@ -434,10 +439,15 @@ Please report these packages to https://github.com/JuliaLang/Homebrew.jl:
     for dep in sorted_deps
         install_and_link(dep; verbose=verbose)
     end
+
+    # Cleanup translated formula once we're done, unless asked not to
+    if !keep_translations
+        delete_all_translated_formulae()
+    end
 end
 
-function add(pkg::BrewPkg; verbose::Bool=false)
-    add(fullname(pkg), verbose=verbose)
+function add(pkg::BrewPkg; verbose::Bool=false, keep_translations::Bool=false)
+    add(fullname(pkg); verbose=verbose, keep_translations=keep_translations)
 end
